# pkg/server/api/docs/swagger.yaml
basePath: /
definitions:
...
  FRPSConfig:
    properties:
      domain:
        type: string
      port:
        type: integer
      protocol:
        type: string
    type: object
...
  GitNamespace:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  GitProvider:
    properties:
      baseApiUrl:
        type: string
      id:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
...
  Repository:
    properties:
      branch:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        type: string
      path:
        type: string
      prNumber:
        type: integer
      sha:
        type: string
      source:
        type: string
      url:
        type: string
    type: object
  ServerConfig:
    properties:
      apiPort:
        type: integer
      binariesPath:
        type: string
      frps:
        $ref: '#/definitions/FRPSConfig'
      gitProviders:
        items:
          $ref: '#/definitions/GitProvider'
        type: array
      headscalePort:
        type: integer
      id:
        type: string
      providersDir:
        type: string
      registryUrl:
        type: string
      serverDownloadUrl:
        type: string
      targetsFilePath:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Daytona Server API
  title: Daytona Server API
  version: 0.1.0
paths:
  /gitprovider:
    get:
      description: List git providers
      operationId: ListGitProviders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/GitProvider'
            type: array
      summary: List git providers
      tags:
      - gitProvider
    post:
      description: Set Git provider
      operationId: SetGitProvider
      parameters:
      - description: Git provider
        in: body
        name: gitProviderData
        required: true
        schema:
          $ref: '#/definitions/GitProvider'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Set Git provider
      tags:
      - gitProvider
  /gitprovider/{gitProviderId}:
    delete:
      description: Remove Git provider
      operationId: RemoveGitProvider
      parameters:
      - description: Git provider
        in: path
        name: gitProviderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Remove Git provider
      tags:
      - gitProvider
    get:
      description: Get Git provider
      operationId: GetGitProvider
      parameters:
      - description: Git provider
        in: path
        name: gitProviderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GitProvider'
      summary: Get Git provider
      tags:
      - gitProvider
  /gitprovider/{gitProviderId}/{namespaceId}/repositories:
    get:
      description: Get Git repositories
      operationId: GetRepositories
      parameters:
      - description: Git provider
        in: path
        name: gitProviderId
        required: true
        type: string
      - description: Namespace
        in: path
        name: namespaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Repository'
            type: array
      summary: Get Git repositories
      tags:
      - gitProvider
  /gitprovider/{gitProviderId}/namespaces:
    get:
      description: Get Git namespaces
      operationId: GetNamespaces
      parameters:
      - description: Git provider
        in: path
        name: gitProviderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/GitNamespace'
            type: array
      summary: Get Git namespaces
      tags:
      - gitProvider
...
  /gitprovider/context/{gitUrl}:
    get:
      description: Get Git context
      operationId: GetGitContext
      parameters:
      - description: Git URL
        in: path
        name: gitUrl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Repository'
      summary: Get Git context
      tags:
      - gitProvider
...
  /server/config:
    get:
      description: Get the server configuration
      operationId: GetConfig
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServerConfig'
      summary: Get the server configuration
      tags:
      - server
    post:
      consumes:
      - application/json
      description: Set the server configuration
      operationId: SetConfig
      parameters:
      - description: Server configuration
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/ServerConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ServerConfig'
      summary: Set the server configuration
      tags:
      - server
...
schemes:
- http
swagger: "2.0"