# pkg/serverapiclient/model_git_namespace.go
/*
Daytona Server API

Daytona Server API

API version: 0.1.0
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package serverapiclient

import (
	"encoding/json"
)

// checks if the GitNamespace type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &GitNamespace{}

// GitNamespace struct for GitNamespace
type GitNamespace struct {
	Id *string `json:"id,omitempty"`
	Name *string `json:"name,omitempty"`
}

// NewGitNamespace instantiates a new GitNamespace object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewGitNamespace() *GitNamespace {
	this := GitNamespace{}
	return &this
}

// NewGitNamespaceWithDefaults instantiates a new GitNamespace object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewGitNamespaceWithDefaults() *GitNamespace {
	this := GitNamespace{}
	return &this
}

// GetId returns the Id field value if set, zero value otherwise.
func (o *GitNamespace) GetId() string {
	if o == nil || IsNil(o.Id) {
		var ret string
		return ret
	}
	return *o.Id
}

// GetIdOk returns a tuple with the Id field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *GitNamespace) GetIdOk() (*string, bool) {
	if o == nil || IsNil(o.Id) {
		return nil, false
	}
	return o.Id, true
}

// HasId returns a boolean if a field has been set.
func (o *GitNamespace) HasId() bool {
	if o != nil && !IsNil(o.Id) {
		return true
	}

	return false
}

// SetId gets a reference to the given string and assigns it to the Id field.
func (o *GitNamespace) SetId(v string) {
	o.Id = &v
}

// GetName returns the Name field value if set, zero value otherwise.
func (o *GitNamespace) GetName() string {
	if o == nil || IsNil(o.Name) {
		var ret string
		return ret
	}
	return *o.Name
}

// GetNameOk returns a tuple with the Name field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *GitNamespace) GetNameOk() (*string, bool) {
	if o == nil || IsNil(o.Name) {
		return nil, false
	}
	return o.Name, true
}

// HasName returns a boolean if a field has been set.
func (o *GitNamespace) HasName() bool {
	if o != nil && !IsNil(o.Name) {
		return true
	}

	return false
}

// SetName gets a reference to the given string and assigns it to the Name field.
func (o *GitNamespace) SetName(v string) {
	o.Name = &v
}

func (o GitNamespace) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o GitNamespace) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	if !IsNil(o.Id) {
		toSerialize["id"] = o.Id
	}
	if !IsNil(o.Name) {
		toSerialize["name"] = o.Name
	}
	return toSerialize, nil
}

type NullableGitNamespace struct {
	value *GitNamespace
	isSet bool
}

func (v NullableGitNamespace) Get() *GitNamespace {
	return v.value
}

func (v *NullableGitNamespace) Set(val *GitNamespace) {
	v.value = val
	v.isSet = true
}

func (v NullableGitNamespace) IsSet() bool {
	return v.isSet
}

func (v *NullableGitNamespace) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableGitNamespace(val *GitNamespace) *NullableGitNamespace {
	return &NullableGitNamespace{value: val, isSet: true}
}

func (v NullableGitNamespace) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableGitNamespace) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}